Tidied up code
Separated code based on which tab page it belong to, for legibility. Removed old / redundant code.
